# specs/openiap-kit/generated/openapi/commerce-protocol.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenIAP Commerce Protocol",
    "version": "1.0",
    "description": "REST binding of the OpenIAP Commerce Protocol operation surface. Generated from commerce-protocol.graphql — do not edit. Error statuses follow the manifest's errorStatus table; every error body is a ProtocolErrorResponse.",
    "license": {
      "name": "MIT"
    }
  },
  "paths": {
...
    "/commerce/v1/subscriptions/status": {
      "get": {
        "operationId": "subscriptionStatus",
        "summary": "A developer backend reads one authenticated user's subscription standing",
        "description": "A developer backend reads one authenticated user's subscription standing. Server role only: a shipped app must not be able to walk arbitrary user identities, which is why the verification role is refused here.",
        "tags": [
          "entitlements"
        ],
        "security": [
          {
            "serverCredential": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionStatusResult"
                }
              }
            }
          },
          "400": {
            "description": "`INVALID_REQUEST`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "`UNAUTHORIZED`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "`FORBIDDEN`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "`RATE_LIMITED`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "`INTERNAL_ERROR`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "description": "The opaque app-scoped user identity the caller's backend selected, in the identity space shared by provider and caller.",
              "$ref": "#/components/schemas/Identifier"
            }
          }
        ]
      }
    },
    "/commerce/v1/entitlements": {
      "get": {
        "operationId": "entitlements",
        "summary": "A developer backend reads one authenticated user's access decision",
        "description": "A developer backend reads one authenticated user's access decision. Server role only, and fail-close: anything the provider cannot classify as currently entitled is absent from the answer.",
        "tags": [
          "entitlements"
        ],
        "security": [
          {
            "serverCredential": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntitlementsResult"
                }
              }
            }
          },
          "400": {
            "description": "`INVALID_REQUEST`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "`UNAUTHORIZED`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "`FORBIDDEN`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "`RATE_LIMITED`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "`INTERNAL_ERROR`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtocolErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "description": "The opaque app-scoped user identity the caller's backend selected, in the identity space shared by provider and caller.",
              "$ref": "#/components/schemas/Identifier"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SubscriptionStatusResult": {
        "description": "The user's current subscription standing. `active` answers the gate for the user as a whole; `subscription` is the most relevant record — the current entitling subscription when one exists, otherwise the provider's most recent record as context. It is omitted when the provider has no record for this user. A provider that cannot enumerate the user's records completely MUST fail the operation instead of answering from a partial read.",
        "type": "object",
        "required": [
          "active"
        ],
        "additionalProperties": false,
        "properties": {
          "active": {
            "description": "Whether the user is entitled to anything right now. The gate for simple access checks.",
            "type": "boolean"
          },
          "subscription": {
            "$ref": "#/components/schemas/SubscriptionStatusSnapshot"
          }
        }
      },
...
      "EntitlementsResult": {
        "description": "Every product the user may access right now, with the entitling subscription records. Unknown, expired, and ambiguous records contribute nothing: a product appears only when its gate is open at the provider's read time. A provider that cannot enumerate the user's records completely MUST fail the operation instead of answering from a partial read.",
        "type": "object",
        "required": [
          "userId",
          "productIds",
          "subscriptions"
        ],
        "additionalProperties": false,
        "properties": {
          "userId": {
            "$ref": "#/components/schemas/Identifier"
          },
          "productIds": {
            "description": "Every product the user may access right now, deduplicated.",
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1
            },
            "uniqueItems": true
          },
          "subscriptions": {
            "description": "The subscription records whose open gates produced `productIds`.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SubscriptionStatusSnapshot"
            }
          }
        }
      },
...
      "ProtocolErrorResponse": {
        "description": "The REST error envelope. Every non-success REST response carries exactly this shape — CLOSED, so an error body cannot smuggle members past the tokenless rules (a failure response is the easiest place to hide a leak, because callers rarely inspect it).",
        "type": "object",
        "required": [
          "error"
        ],
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ProtocolError"
          }
        }
      },
...
      "Identifier": {
        "description": "An opaque, implementation-assigned identifier. Consumers MUST treat it as an opaque string and MUST NOT parse structure out of it.",
        "type": "string",
        "minLength": 1,
        "maxLength": 256
      },
      "SubscriptionStatusSnapshot": {
        "description": "One subscription as the provider currently records it, shaped for a server-side read. It is tokenless by construction: this object is CLOSED, so the schema itself rejects any member beyond those listed — no purchase token, store transaction identity, signed receipt, or provider-internal record identifier can appear, and the response is safe to hold in a developer backend.",
        "type": "object",
        "required": [
          "productId",
          "state",
          "active"
        ],
        "additionalProperties": false,
        "properties": {
          "productId": {
            "description": "Store product identifier of the subscription.",
            "type": "string",
            "minLength": 1
          },
          "state": {
            "$ref": "#/components/schemas/SubscriptionState"
          },
          "active": {
            "description": "The entitlement gate for this subscription, evaluated with the predicate in SPEC.md 2.3 at the provider's read time. A caller gates access on this member, never on `state`.",
            "type": "boolean"
          },
          "store": {
            "$ref": "#/components/schemas/Store"
          },
          "expiresAt": {
            "$ref": "#/components/schemas/Timestamp"
          },
          "renewsAt": {
            "$ref": "#/components/schemas/Timestamp"
          },
          "willRenew": {
            "description": "Whether the store will attempt another billing period, when the provider records it.",
            "type": "boolean"
          },
          "cancellationReason": {
            "description": "Why the subscription stopped renewing, as the same open normalized token space the event envelope uses. Advisory, never a billing fact.",
            "type": "string",
            "minLength": 1
          },
          "startedAt": {
            "$ref": "#/components/schemas/Timestamp"
          },
          "updatedAt": {
            "$ref": "#/components/schemas/Timestamp"
          }
        }
      },
...
      "ProtocolError": {
        "description": "One protocol operation failure. Both bindings carry the same code space: REST wraps it in ProtocolErrorResponse, GraphQL carries `code` in `errors[].extensions`. A message is human-readable and MUST NOT contain credentials, store evidence, stack traces, or implementation source paths.",
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "additionalProperties": false,
        "properties": {
          "code": {
            "description": "A portable protocol error code. The value space is OPEN: a caller MUST tolerate a code it does not recognise and treat it as a failure of the operation, and a MINOR version can add a code. The listed values are the ones this version names; SPEC.md 8 defines each one and its HTTP status.",
            "type": "string",
            "pattern": "^[A-Z][A-Z0-9_]*$",
            "examples": [
              "INVALID_REQUEST",
              "UNAUTHORIZED",
              "FORBIDDEN",
              "NOT_FOUND",
              "PURCHASE_NOT_FOUND",
              "VERIFICATION_FAILED",
              "CONFLICT",
              "RATE_LIMITED",
              "UNSUPPORTED_PROFILE",
              "UNSUPPORTED_STORE",
              "INTERNAL_ERROR"
            ]
          },
          "message": {
            "description": "Human-readable failure summary. Never carries credentials, store evidence, stack traces, or source paths.",
            "type": "string",
            "minLength": 1
          }
        }
      },
...
      "SubscriptionState": {
        "description": "The lifecycle state of a subscription. This is a state, not an event, and it does not by itself decide entitlement — see Entitlement in SPEC.md. These members are PascalCase while event types are lowercase-dotted; the inconsistency is inherited from the deployed 1.0 wire format and is recorded in CONVENTION.md rather than silently corrected, because changing it would break receivers already decoding it.",
        "type": "string",
        "enum": [
          "Active",
          "InGracePeriod",
          "InBillingRetry",
          "Paused",
          "Expired",
          "Revoked",
          "Refunded",
          "Unknown"
        ]
      },
...
    },
    "securitySchemes": {
      "verificationCredential": {
        "type": "http",
        "scheme": "bearer",
        "description": "Verification-role credential: limited to purchase-evidence verification, no account access."
      },
      "serverCredential": {
        "type": "http",
        "scheme": "bearer",
        "description": "Server-role credential for an authenticated developer backend. Never ship it in an app."
      }
    }
  },
...
}